+++ src/nexus_data/__init__.py
@@ -1,0 +1,1 @@
+"""Point-in-time data acquisition for NEXUS-MFP."""